% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_website.R
\name{build_website}
\alias{build_website}
\title{Create LDA viewer Site}
\usage{
build_website(json, odir, ldavis = "", info = "")
}
\arguments{
\item{odir}{A string containing the path to create the website in.}

\item{ldavis}{**optional A string containing the json data generated with createJSON from the LDAvispackage.}

\item{info}{**optional** A string that you want to place in index.html to record some details
about the model.}

\item{viewerJSON}{A string containing the json data built with build_website_json}
}
\value{
A string that is the path to the directory containing the site
}
\description{
The function that creates the ldaviewer website.
Reads in the json data from build_website_json and writes
it to the output directory along with the html/js/css from /inst/.
Optionally can be passed the lda.json data created by 
createJSON from ldavis. If so, will add that page to the site.
}
\examples{
\dontrun{
build_website(viewerjson, "./site/")
build_website(viewerjson, "./site/", ldavis=myjson)
build_website(viewerjson, "./site/", info="alpha=.05 beta=0.1")
}
}
