% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dt.R, R/create_td.R
\name{create_dt}
\alias{create_dt}
\title{Create Doc Topics JSON}
\usage{
create_dt(x, count = 15)

create_dt(x, count = 15)
}
\arguments{
\item{x}{A doc topic matrix.}

\item{count}{**optional** Number of topics per document to store, default=15)}

\item{fnames}{A character vector of filenames with length = nrows(x)}

\item{x}{A doc topic matrix.}

\item{count}{**optional** Number of documents per document to store, default=15)}
}
\value{
A string containing the JSON.

A string containing the JSON.
}
\description{
A function to create a string containing two arrays. First array
has the top topics per doc. Second has the scores for those topics.
If doc topics is incredibly large, then just pass a sample.
Output string has this format: 
var inds = [11,23...];
var scores = [0.04,0.02,...];

A function to create a string containing one array which has
has the top docs per topic. 
If doc topics is incredibly large, then just pass a sample.
Output string has this format: 
var td_inds = [11,23...];
}
\examples{
\dontrun{
dt_small = create_dt(dt) 
}
\dontrun{
td_small = create_td(dt) 
td_small = create_td(dt, 20) 
}
}
