% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_contents.R
\name{create_file_contents}
\alias{create_file_contents}
\title{Create File Contents JSON}
\usage{
create_file_contents(files, nchars = 3000)
}
\arguments{
\item{files}{A string vector of filenames}

\item{nchars}{**optional** Number of characters per document to store.}
}
\value{
A string containing the JSON.
}
\description{
A function to create a string containing the first X characters
of each file given a vector of filenames. If for whatever reason
the file can't be read from (doesn't exist or lack of permissions)
the string for that file will be blank. 
Output string has this format: 
var file_contents = ["file 1 contents", "more text here", ...];
}
\examples{
\dontrun{
file_contents = create_file_contents("./texts/")
file_contents = create_file_contents("./texts/", 1000)
}
}
