% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_td.R
\name{create_td}
\alias{create_td}
\title{Create Small Topic Docs JSON}
\usage{
create_td(x, fnames, count = 15)
}
\arguments{
\item{x}{A doc topic matrix.}

\item{fnames}{A character vector of filenames with length = nrows(x)}

\item{count}{**optional** Number of documents per topic to store, default=15)}
}
\value{
A string containing the JSON.
}
\description{
A function to create a string containing the topic docs array.
Takes as input the doc topics array and the names of the files.
If doc topics is incredibly large, then just pass a sample
to this punction and the filenames for that sample.
Output string has this format: var topicdocs = [
{'name':1, 'scores':[0.04,...], 'docs':['filename1','filename2'...]},
{...}
]
}
\examples{
\dontrun{
td_small = create_td(dt, fnames)
td_small = create_td(dt, fnames, 20)
}
}
