% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tt.R
\name{create_tt}
\alias{create_tt}
\title{Create Topic Terms JSON}
\usage{
create_tt(x, vocab, count = 200)
}
\arguments{
\item{x}{A topic term matrix.}

\item{count}{**optional** Number of terms per topic to store, default=200)}
}
\value{
A string containing the JSON.
}
\description{
A function to create a string containing the topic terms array.
Takes as input the topic terms array.
Output string has this format:
var inds = [21000,131,...];
var scores = [0.003, 0.002,...];
}
\examples{
\dontrun{
tt_small = create_tt(tt)
tt_small = create_tt(tt, 100)
}
}
