% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_viewer.R
\name{create_viewer}
\alias{create_viewer}
\title{Create LDA viewer Site}
\usage{
create_viewer(dt, tt, vocab, fnames, textpath, odir, ldavisJSON = "",
  metainfo = "", verbose = FALSE)
}
\arguments{
\item{dt}{A doc topic matrix.}

\item{tt}{A topic terms matrix.}

\item{vocab}{A vector of the vocab length = ncols(tt)}

\item{fnames}{A vector of the filenames length = nrows(dt)}

\item{textpath}{A string containing the path to the directory containing
all the text files.}

\item{odir}{A string containing the path to create the website in.}

\item{ldavisJSON}{**optional A string containing the json data generated with createJSON from the LDAvispackage.}

\item{metainfo}{**optional** A string that you want to place in index.html to record some details
about the model.}

\item{verbose}{**optional** A bool, displays some extra prints if set to TRUE, default=FALSE.}
}
\value{
A string that is the path to the directory containing the site
}
\description{
The function that creates the ldaviewer website.
Reads in the doc topics and topic terms array,
runs them through the appropriate functions to create
the js files necessary for the site. Moves those files and
the files in /inst/ to the output directory.
Optionally can be passed the lda.json data created by 
createJSON from ldavis. If so, will add that page to the site.
}
\examples{
\dontrun{
create_viewer(dt,tt,v,f, "./texts", "./site/")
create_viewer(dt,tt,v,f, "./texts", "./site/", ldavisJSON=myjson)
create_viewer(dt,tt,v,f, "./texts", "./site/", metainfo="alpha=.05 beta=0.1")
create_viewer(dt,tt,v,f, "./texts", "./site/", verbose=TRUE)
}
}
