% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_viewer_json.R
\name{create_viewer_json}
\alias{create_viewer_json}
\title{Create LDA viewer JSON}
\usage{
create_viewer_json(tt, dt, vocab, fnames, texts, verbose = FALSE)
}
\arguments{
\item{tt}{A topic terms matrix.}

\item{dt}{A doc topic matrix.}

\item{vocab}{A vector of the vocab length = ncols(tt)}

\item{fnames}{A vector of the filenames length = nrows(dt) and length(texts)}

\item{texts}{A character vector of the file contents. Can be genereated with the load_texts() function.}

\item{verbose}{**optional** A bool, displays some extra prints if set to TRUE, default=FALSE.}
}
\value{
A string containing the complete json for ldaviewer
}
\description{
Creates all the json data necessary for the ldaviewer site.
Reads in doc topics, topic terms, the file contents, filenames
and vocab.
}
\examples{
\dontrun{
myjson = create_viewer_json(tt,dt,v,f, docs)
myjson = create_viewer_json(tt,dt,v,f, docs, verbose=TRUE)
}
}
