% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_texts.R
\name{load_texts}
\alias{load_texts}
\title{Load Texts}
\usage{
load_texts(files, nchars = 3000)
}
\arguments{
\item{files}{A string vector of filenames or a single file name.}

\item{nchars}{**optional** Number of characters per document to store.}
}
\value{
A character vector of the first nchars from each file
}
\description{
A function to create a vector containing the first X characters
of each document given a vector of filenames. If a single file is passed
then the function assumes each line has a different document.
If for whatever reason the file can't be read from (doesn't exist or lack of permissions)
the string for that doc will be blank.
}
\examples{
\dontrun{
file_contents = load_texts(files)
file_contents = load_texts(files, 1000)
}
}
