% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_td.R
\name{create_td}
\alias{create_td}
\title{Create Topics Docs JSON}
\usage{
create_td(x, count = 15)
}
\arguments{
\item{x}{A doc topic matrix.}

\item{count}{**optional** Number of documents per document to store, default=15)}
}
\value{
A string containing the JSON.
}
\description{
A function to create a string containing one array which has
has the top docs per topic. 
If doc topics is incredibly large, then just pass a sample.
Output string has this format: 
var td_inds = [11,23...];
}
\examples{
\dontrun{
td_small = create_td(dt) 
td_small = create_td(dt, 20) 
}
}
