% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tt.R
\name{create_tt}
\alias{create_tt}
\title{Create Small Topic Terms JSON}
\usage{
create_tt(x, vocab, count = 200)
}
\arguments{
\item{x}{A topic term matrix.}

\item{count}{**optional** Number of terms per topic to store, default=200)}

\item{fnames}{A character vector of vocab with length = ncols(x)}
}
\value{
A string containing the JSON.
}
\description{
A function to create a string containing the topic terms array.
Takes as input the topic terms array and the vocab.
Output string has this format: var topicterms = [
{'name':1, 'scores':[0.04,...], 'terms':['word1", "word2",...]},
{...}
]
}
\examples{
\dontrun{
dt_small = create_dt(dt, vocab)
dt_small = create_dt(dt, vocab, 100)
}
}
